device pins: Create classes that can be something other than a Pin subclass
